Refactor SlWiFiDriver to improve handling of SSID and PSK lengths. Updated methods to use size_t for length variables and adjusted memcpy calls for better type safety. Enhanced consistency in WiFiCredentials usage across the implementation.

+++ src/platform/silabs/wifi/SiWx/WifiInterfaceImpl.cpp
@@ -473,8 +473,8 @@ sl_status_t SetWifiConfigurations()
         .config = {
             .ssid = {
                 .value  = { 0 },
-                //static cast because the types dont match
-                .length = wfx_rsi.credentials.ssidLen,
+                // static cast because the types dont match
+                .length = static_cast<uint8_t>(wfx_rsi.credentials.ssidLen),
             },
             .channel = {
                 .channel = SL_WIFI_AUTO_CHANNEL,
